Load data and apply projections in spatial chunks

Automatically load and process climate model projections in a memory efficient way. Useful for analysing large areas

load_data_and_projections(
  variable,
  years.hist = NULL,
  years.proj,
  path.to.data,
  path.to.obs = NULL,
  xlim,
  ylim,
  aggr.m = "none",
  chunk.size,
  overlap = 0.25,
  season,
  lowert = NULL,
  uppert = NULL,
  consecutive = F,
  duration = "max",
  bias.correction = F,
  domain = NULL,
  n.sessions = 1
)

Arguments

variable
character, indicating variable name

years.hist
Numerical range, years to select for historical simulations and observations

years.proj
Numerical range, years to select for projections

path.to.data
path to the directory containing the RCP/SSPs folders and historical simulations (optional). For example, home/user/data/. data would contain subfolders with the climate/impact models. Historical simulations have to be contained in a folder called historical. If path.to.data is set as CORDEX-CORE, CORDEX-CORE simulations from RCM RegCM4 will be loaded

path.to.obs
Default to NULL, if not, indicate the absolute path to the directory containing a reanalysis dataset, for example ERA5. To automatically load W5E5. specify W5E5

xlim
numeric of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box selected.

ylim
same as xlim, but for the selection of the latitudinal range.

aggr.m
character, monthly aggregation. One of none, mean or sum

chunk.size
numeric, indicating the number of chunks. The smaller the better when working with limited RAM

overlap
numeric, amount of overlap needed to create the composite. Default 0.5

season
numeric, seasons to select. For example, 1:12

lowert
numeric of length=1, lower threshold

uppert
numeric of length=1, upper threshold

consecutive
logical, to use in conjunction with lowert or uppert

duration
character, either "max" or "total"

bias.correction
logical

domain
specify the CORDEX-CORE domain (e.g AFR-22, EAS-22). Used with path.to.data = CORDEX-CORE. Default is NULL

n.sessions
numeric, number of sessions to use, default is one. Parallelisation can be useful when multiple scenarios are used (RCPS, SSPs). However, note that parallelising will increase RAM usage

Value

list with SpatRaster. .[[1]] contains SpatRaster for the ensemble mean. .[[2]] contains SpatRaster for the ensemble sd and .[[3]] conins SpatRaster for individual models