Subset years in CAVAlist

Select specific years for projections or observations in a CAVAlist

Usage

years_selection(CAVAlist, years, projections = T)

Arguments

CAVAlist

output of the load_data or load_data_hub function

years

numeric, years to select

projections

logical, whether years selection should be applied to projections or observations